+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.0.1
 
-- Adding `computed`, `effect`, `signal`
+- Adding `computed`, `effect`, `signal`, `untracked`
 - Adding tests
 - Adding web example